+++ src/App.jsx
@@ -4,7 +4,7 @@ import List from './pages/List';
 import Paper from './pages/Paper';
 import Message from './pages/Message';
 import NavLayout from './layout/NavLayout';
-import Post from './pages/PostPage';
+import Post from './pages/Post';
 
 const App = () => {
   return (
