+++ src/pages/Post/components/CheckMark/CheckMarkStyled.js
